some JavaScript files renamed
